pyproject.toml
[project]
name = "threestudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.8.0",
    "bitsandbytes==0.38.1",
    "clip-openai>=1.0.post20230121",
    "controlnet-aux>=0.0.10",
    "diffrp-nvdiffrast>=0.3.3.1",
    "diffusers<0.20",
    "einops>=0.8.1",
    "envlight>=0.1.0",
    "gradio>=5.1.0",
    "huggingface-hub>=0.33.0",
    "imageio[ffmpeg]>=2.28.0",
    "jaxtyping>=0.3.2",
    "kornia>=0.8.1",
    "libigl>=2.6.1",
    "lightning==2.0.0",
    "matplotlib>=3.10.3",
    "nerfacc==0.5.2",
    "networkx>=3.5",
    "numpy>=2.3.0",
    "omegaconf==2.3.0",
    "opencv-python>=4.11.0.86",
    "pybind11>=2.13.6",
    "pymcubes>=0.1.6",
    "pysdf>=0.1.9",
    "safetensors>=0.5.3",
    "sentencepiece>=0.2.0",
    "setuptools-rust>=1.11.1",
    "taming-transformers-rom1504>=0.0.6",
    "tensorboard>=2.19.0",
    "tokenizers>=0.21",
    "torch==2.7.0",
    "torchmetrics>=1.7.3",
    "transformers>=4.52.4",
    "trimesh[easy]>=4.6.12",
    "typeguard>=4.4.4",
    "wandb>=0.20.1",
    "xatlas>=0.0.10",
    "xformers>=0.0.30",
]

[tool.uv]
# Resolve for macOS, but not for Linux or Windows.
environments = [
  "platform_machine == 'x86_64' and sys_platform == 'linux'"
]
